+++ gists/kerchunk.py
@@ -27,11 +27,6 @@ from mbo_utilities.lazy_array import imread, imwrite
 import uuid
 import subprocess
 
-def _get_git_commit() -> str:
-    try:
-        return subprocess.check_output(["git", "rev-parse", "HEAD"]).decode().strip()
-    except Exception:
-        return "unknown"
 
 def dump_kerchunk_reference(tif_path: Path, base_dir: Path) -> dict:
     """
@@ -133,75 +128,7 @@ def load_zarr_data(spec_path):
     z_arr = zarr.open(store, mode="r",)
     return z_arr
 
-def _load_existing(save_path: Path) -> list[dict[str, Any]]:
-    if not save_path.exists():
-        return []
-    try:
-        return json.loads(save_path.read_text())
-    except Exception:
-        return []
-
-def _increment_label(existing: list[dict[str, Any]], base_label: str) -> str:
-    count = 1
-    labels = {e["label"] for e in existing if "label" in e}
-    if base_label not in labels:
-        return base_label
-    while f"{base_label} [{count+1}]" in labels:
-        count += 1
-    return f"{base_label} [{count+1}]"
-
-
-def _benchmark_indexing(
-    arrays: dict[str, np.ndarray | da.Array | zarr.Array],
-    save_path: Path,
-    num_repeats: int = 5,
-    index_slices: dict[str, tuple[slice | int, ...]] = None,
-    label: str = None,
-):
-    if index_slices is None:
-        index_slices = {
-            "[:200,0,:,:]": (slice(0, 200), 0, slice(None), slice(None)),
-            "[:,0,:40,:40]": (slice(None), 0, slice(0, 40), slice(0, 40)),
-        }
-
-    results = {}
-    for name, array in arrays.items():
-        results[name] = {}
-        for label_idx, idx in index_slices.items():
-            times = []
-            for _ in range(num_repeats):
-                t0 = time.perf_counter()
-                val = array[idx]
-                if isinstance(val, da.Array):
-                    val.compute()
-                elif hasattr(val, "read"):
-                    np.array(val)
-                t1 = time.perf_counter()
-                times.append(t1 - t0)
-            results[name][label_idx] = {
-                "min": round(min(times), 3),
-                "max": round(max(times), 3),
-                "mean": round(sum(times) / len(times), 3),
-            }
-
-    save_path = Path(save_path)
-    save_path.parent.mkdir(parents=True, exist_ok=True)
-
-    existing = _load_existing(save_path)
-    final_label = _increment_label(existing, label or "Unnamed Run")
-
-    entry = {
-        "uuid": str(uuid.uuid4()),
-        "git_commit": _get_git_commit(),
-        "label": final_label,
-        "index_slices": list(index_slices.keys()),
-        "results": results,
-    }
-
-    existing.append(entry)
-    save_path.write_text(json.dumps(existing, indent=2))
-
-    return results
+s
 
 if __name__ == "__main__":
     data = imread(
